[build-properties] Allow setting kotlin gradle plugin version (#36763)
# Why

It is currently not possible to use kotlin 2+ with expo sdk 52. This
fixes it. We currently have a hard requirement of kotlin 2+ for the
android stripe sdk so we need to support it in the react-native sdk. A
lot of users are still on expo 52 so it would be good to support it.

Note that this must only be used for sdk 52, in sdk 53 the kotlin
version is no longer specified the same way and is not compatible with
this.

# How

This passes the kotlin version explicitly to the kotlin gradle plugin.
This allows overriding what the react-native gradle plugin sets.