[docs] update font docs for clarity (#36725)
# Why

Improve the documentation for `expo-font` to clarify the two approaches
for adding custom fonts to an app: using the config plugin (recommended)
or loading fonts at runtime. It was previously easy to have a quick look
at the docs and end up with the impression that you need to use the
config plugin as well as `useFonts`.

closes https://github.com/expo/expo/issues/33673

I will duplicate this to SDK 53 docs once we agree on the copy.

# How

- Clarified the documentation to explicitly mention both font loading
approaches
- Added more context about when to use the config plugin vs runtime
loading
- Added a description for the  `useFonts` hook example
- Specified that `loadAsync` is a "runtime method" to distinguish it
from the build-time config plugin approach

# Test Plan

- locally

# Checklist

- [x] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)

---------

Co-authored-by: Aman Mittal <amandeepmittal@live.com>