fix(babel-preset-expo): Disable `minifyTypeofWindow` for web targets temporarily (#36773)
# Why

Resolves #36620

Due to `asyncType` being used to communicate whether we're bundling for
a Web Worker, it's currently tricky to mark the transform step as being
for Web Workers. It's also unclear whether passing this on somehow would
break Metro's cache key for the affected files.

However, `minifyTypeofWindow` makes the incorrect assumption of `typeof
window === 'object'` for web workers.

# How

Temporarily disable `minifyTypeofWindow` for all web targets to prevent
this from becoming a problem, until we have a better solution.

# Test Plan

- Manually tested