[go] Fix `isEdgeToEdge` in Expo Go (#36829)
# Why

`RNEdgeToEdge` package wasn't being registered in expo-go native modules
list, which broke `isEdgeToEdge` method from
`react-native-edge-to-edge`. Because of this `expo-status-bar` and
`expo-navigation-bar` wouldn't use `SystemBars` implementation from
`react-native-edge-to-edge` breaking the status and navigation bar (Expo
Go is always edge-to-edge).

# How

Added the `react-native-edge-to-edge` as an dependency and added the
edge-to-edge package to modules in `ExponentPackage.kt`

# Test Plan

Tested in Expo Go on Android