[router] Fix useLocalSearchParams returning "undefined" for deleted params (#36811)
# Why

Fix: #36791

# How

React Navigation cannot delete params completely, they will still be
tracked as `undefined`. When `useLocalSearchParams` gets the params, it
first decodes them, which causes the value to be exposed as
`'undefined'`.

As a side point, I'm not sure that we need to decode the params any more
in v5.

# Test Plan

Added unit test