[android][kotlin] removed use of internal java member
Meta is rewriting all its java files to Kotlin meaning that we're not able to access private members of classes by subclassing and exposing them any more.

This commit fixes an issue where we exposed the underlying ReadableMap in the ReactStylesDiffMap class by subclassing it. The commit refactors the code to use the ReactStylesDiffMap and its publicly available members.