[expo-go][ios] fix barometer resolution (#26345)
# Why
Fixes a rounding of the Barometric pressure on Expo Go.
This was previously fixed in the sensors package but not in Expo Go.
https://github.com/expo/expo/pull/9441

# How

By not converting the value to integer, we do the same as the sensors
package and use a double.
https://github.com/expo/expo/pull/9441

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [x] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [x] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [x] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Vojtech Novak <vonovak@gmail.com>