revert: Revert "noop non-fabric renderer when new architecture is enabled." #36757 (#36913)
# Why

The PR #36757 removes the paper renderer from the bundler when using the
new arch to save space. After testing for a bit we found that
`react-native-gesture-handler` uses `findNodeHandle` which in turn
causes it to import the paper renderer even though we're on the new
arch.

We decided to revert this one.

@EvanBacon 
@Kudo 

# How

This reverts commit 959e44addb78f7f6166ec2112888d024a66758f2.

# Test Plan

Tests green

# Checklist

- [x] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).