fix(config-plugins): Add `package.json:types` and `package.json:exports` fields (#36774)
# Why

We had minimal `package.json` entries on `@expo/config-plugins`, which
can cause problems with some configurations of TypeScript, with it being
potentially unable to resolve types for the package.

# How

Add `package.json:types` field, which was previously missing, and add
`package.json:exports` fields while we're at it.

# Test Plan

- Covered in local type checks with `tsc`
- Create a minimal project with `expo/tsconfig.base` in `tsconfig.json`
as a preset, add `@expo/config-plugins` dependency, import from a nested
file