[navigation-bar] Fix docs inconsistencies (#36966)
# Why

Expo navigation bar has some docs inconsistencies.

- `Using it in edge-to-edge app may lead to unexpected behavior.` makes
it sound like the method can be used in edge-to-edge. It can't. It's a
leftover from back when we assumed that we would allow people to try to
use them in edge-to-edge
- `getVisibilityAsync` is marked as available in edge-to-edge, but will
throw a warning when used (unnecessarily)


# How

Change ` * > This method is supported only when edge-to-edge is
disabled. Using it in edge-to-edge app may lead to unexpected behavior.`
to * > This method is supported only when edge-to-edge is disabled.`
Make `getVisibilityAsync` available in all contexts.

# Test Plan

Tested in BareExpo on Android SDK 35 emulator.