[image] Add imperative api to lock/unlock/reload resource (#36912)
# Why

Animating size (especially width/height) on Image may introduce
unnecessary reloads.

# How

Add an imperative API for advanced use cases where the user needs to
temporarily prevent an image from reloading when the frame changes. This
can be done using `lockResourceAsync` to lock the resource and
`unlockResourceAsync` to release it when it's safe to reload. Since
unlocking does not necessarily trigger a reload, there is also a
separate method called `reloadAsync` that forces reload (even if lock is
active).

# Test Plan

I've added an **Imperative API** screen to NCL