[video][android] Fix video being blank for some sources (#37009)
# Why

Fixes https://github.com/expo/expo/issues/36942

Fixes`onFirstFrameRender` not being emitted for sources with
`pixelWidthHeightRatio` different than 1. Without this event we never
show the video surface.

# How

Include the `pixelWidthHeightRatio` in the calculations for the
`onFirstFrameRender` event.

While I was there I also increased the `epsilon` param to `0.05`, turns
out that 0.01 may be a bit too low for very low resolution videos (I
didn't find any videos that actually break because of this, but I prefer
to stay on the safe side).

# Test Plan

Tested in BareExpo on Android 15 emulator