[video][android] Fix `useExoShutter` prop not being exposed to the JS side. (#37012)
# Why

The `useExoShutter` prop wasn't included in the native module
definition, which means that it didn't do anything.

# How

Added the prop to the view definition. Also since we already shipped a
version of video with the shutter always off, and in my opinion our
implementation of the shutter is better than the original and works the
same as iOS I've changed the default to `false`.

# Test Plan

Tested in BareExpo on Android