[docs] Add `single` mode to "Get Started" page (#37050)
# Why

All export modes have been supported for a while. We should update the
docs to reflect that the `single` export mode is now supported as well.

# How

- Getting started section
- Rewrite this section to mention `single` without removing it. It's
still beneficial for us to prompt the user to change this export value,
but we should simple mention all available values.
- Introduction section
- Remove mention of static/server; This is easily interpreted as a
limitation and we should just remove it