[core] Reduce Expo modules reliance on legacy NativeModules and add TurboModule support (#36996)
# Why

Some of the expo-modules-core functionality currently requires
NativeModules to hook up the initial connection. And allows expo modules
to be implemented using a NativeModule. It doesn't make sense to have
dependencies on the old architecture at this point, so I'm allowing this
code to work with either NativeModules or TurboModules.

react-native-windows is not supporting NativeModules when running the
new arch, so this change also enabled react-native-windows to provide
expo modules in the future.

# How

Basically, ensureNativeModulesAreInstalled now uses TurboModuleRegistry
to access the 'ExpoModulesCore' module. This means that ExpoModulesCore
can be implmented as either a NativeModule or a TurboModule.

And requireNativeModule will now also find turboModules registered with
the expo module's name, and if found will return them wrapped in a proxy
object that provides the same event emitter API as expo modules use.