[video][ios][1/4] `onStatusChange` audit - Fix player not entering 'error' state when loading fails (#37177)
# Why

The new loading method was catching the loading error and throwing it.
We should actually ignore the error and let the player handle it.

Fixes https://github.com/expo/expo/issues/36673

# How

Ignored the loading error, and assigned the source to the player
anyways. This way the player enters the .error state - the same as with
the old synchronous loading method.

# Test Plan

Tested in BareExpo