[video][ios][2/4] onStatusChange audit - fix status while loading a new source asynchronously (#37180)
# Why

With asynchronous loading the player will continue playing the current
source until the next one has been loading. This means that the status
stays at `.readyToPlay` even though the player is loading a new source.

# How

This PR adds a listener functionality to the `VideoSourceLoader` while
the loader is loading a new source the current status resulting from
playback is ignored and `.loading` is returned.

# Test Plan

Tested on a physical iOS 18 device.