[video][ios][3/4] onStatusChange audit - Fix player going into `loading` status for a single frame when unpausing (#37181)
# Why
Fixes https://github.com/expo/expo/issues/33738

The player will go into `.loading` for a few milliseconds state every
time it is unpaused. This was because the `AVPlayer.timeControlStatus`
will go into a waiting state.

# How

We can check for the reason for the waiting state. When the bug appears
the state is always `.evaluatingBufferingRate`. We can filter it out
without any negative side effects to detecting the loading state.

# Test Plan

Tested in BareExpo on iOS 18 device.