[core][Android] Support object arrays in type converters. (#37276)
# Why

To support object arrays like 
```kt
Function("arrayArrayArrayInt") { a: Array<Array<Int>> -> a }
```

# How

* Extract primitive array converters from `cachedConverters` into
`cachedPrimitiveArrayConverters`.
* Improve `isPrimitiveArray` function to correctly detect primitive
arrays.

It's related to this kotlin issue:
https://youtrack.jetbrains.com/issue/KT-52170/Reflection-typeOfArrayLong-gives-classifier-LongArray

Initially, to check whether an array was a primitive array, I used
`.toString()`. However, to get a string, the entire type had to be
evaluated. I later discovered that primitive arrays have empty
`arguments` and object arrays include the inner type as an argument. Now
we can reliably detect if an array is a primitive array or it's just a
kotlin issue.

# Test Plan

CI should be green ✅