[expo-updates] Fix iOS native debugging for iOS (#37323)
# Why

When built with native debug, an Expo iOS app made from the default expo-router template shows the redbox error

```
[runtime not ready]: Error: Cannot create devtools websocket connections in embedded environments.
```

The only way to prevent this is by setting DEV=false when building the JS bundle.

This PR enforces that for iOS (it is already this way for Android), and fixes an iOS crash that occurs in native debug when the app is run for the first time.

# How

- Add `CONFIGURATION=Release` to Xcode environment for native debug
- Fix the crash in AppLoader.swift

# Test Plan

- Tested with an app created with the default template
- CI should pass

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [ ] I added a `changelog.md` entry and rebuilt the package sources according to [this short guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)