[expo-updates] Fix iOS native debugging for iOS (#37323)
When built with native debug, an Expo iOS app made from the default expo-router template shows the redbox error

```
[runtime not ready]: Error: Cannot create devtools websocket connections in embedded environments.
```

The only way to prevent this is by setting DEV=false when building the JS bundle.

This PR enforces that for iOS (it is already this way for Android), and fixes an iOS crash that occurs in native debug when the app is run for the first time.

- Add `CONFIGURATION=Release` to Xcode environment for native debug
- Fix the crash in AppLoader.swift

- Tested with an app created with the default template
- CI should pass