chore(workflows,ios): update xcode to latest version
When running our workflows we're seeing that iOS builds are failing with the following exception:
https://github.com/expo/expo/runs/43778152626#step:12:58

This error states that we're using a wrong version of XCode (16.0) where a higher one was expected (16.1).

The default XCode version for github actions running on the `macos-15` runner is XCode 16 - which we also set in our workflows.

The runner supports up to XCode 16.4 (latest version):

https://github.com/actions/runner-images/blob/main/images/macos/macos-15-Readme.md#xcode

This commit upgrades the expected XCode version from 16.0 (explicitly) to 16.1 (explicitly) in the Expo workflows that builds with XCode.