[image] Make `useAppleWebpCodec` a view prop instead of the source property (#37300)
# Why

#35802 introduced the `useAppleWebpCodec` property of the image source,
however with this API it's not possible to use it with local assets
(using `require` function) as they are represented by numbers.
Workaround for the user would be to resolve that number to an object
using `resolveAssetSource` from RN, but it's not ideal as in theory it's
an internal function.

# How

I don't see any advantage of having it a source property instead of a
view prop, so I moved that.
I've also rephrased the docs to make its description a bit more clear
and mention known issues with the default codec.