[video][android] Fix aspect ratio when auto-entering PiP (#37225)
# Why

Fixes the aspect ratio always being 16:9 when auto-entering pip on
Android.

# How

Listen for the video source loaded event, get the first video track and
it's aspect ratio. Apply the aspect ratio to the PiP config.

Also did a small refactor of the PiP code to make it a bit more
readable.

# Test Plan

Tested in BareExpo on Android 15 emulator