[ios][audio] Fixed status property bug when track finishes
During the development of an audio player using the new expo-audio package, I noticed on iPhone that the status property "playing" did not change to false when the track finished. Instead, a new property "isPlaying" appeared with the value false. This commit replaces the incorrect key with "playing" for consistency with the rest of the code.