feat(expo): enable async require by default (#36405)
# Why

- Allow more people to use async bundling to speed up their app DX by
default.
- Make metro web fast refresh easier to enable.

# How

- Move async-require and fast refresh code out of metro-require and into
the default Expo side effect. All of the code from `@expo/metro-runtime`
should ultimately live in the expo package as it's required for E2E
functionality.
- Bring over the fast refresh and message socket features for Metro web
and RSC HMR too.
- Introduce a new global `globalThis.__expo_dev_resetErrors` for
connecting the optional error boundary in metro-runtime with the HMR
observer in `expo`. This can be removed in the future when we land the
new error overlay.

# Test Plan

- TBD

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>