[router] create href preview component (#37335)
# Why

To render peek and pop previews, we need the ability to render a different route within the current route. This PR introduces HrefPreview, which allows rendering a preview of any route based on a given href.

# How

Using the react-navigation state, the preview retrieves the component associated with the specified href. This component is then rendered in a "sandboxed" environment where:
- Navigation is disabled.
- usePathname and other hooks return the correct values

# Test Plan

- Unit tests

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [x] I added a `changelog.md` entry and rebuilt the package sources according to [this short guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)