[router] add modal component (#37365)
# Why

Currently, to display a react-native-screens modal, the user needs to create a screen in the stack. This PR simplifies that flow by introducing a new modal component that can be used anywhere in the code.

# How

By wrapping the Expo root in a native stack, the modal can be rendered over all content. Using context, the modal configuration is passed to the stack, and the corresponding screens are rendered.

# Test Plan

1. Manual testing on all platforms
2. Unit tests

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [x] I added a `changelog.md` entry and rebuilt the package sources according to [this short guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)