[CI][expo-updates] dev client updates E2E Maestro test (#37646)
# Why

- There was a script to set up the test project for updates E2E dev
client, but no test was implemented in Detox
- This implements the test in Maestro:
  - Verify state when updates and dev client are both enabled
- Load and run an update from the test update server with the dev client
UI

# Test Plan

CI should pass. Both tests are passing on iOS, but Android will not load
from the update server (investigating).

I have pushed a commit to disable the Android tests that work locally,
but not in Linux workers. Once we have a fix for this, these tests can
be reenabled. The issue with Linux workers is tracked in ENG-16303.