[tools] Throw error when attempting publishing specific packages as canaries. (#37687)
# Why

There is a major bug in the publish canaries script when calling it with package names (it publishes templates with the sdk-53 tag).

This option also won't work due to npm/semver constraints.

# How

Added an error (plus a reminder to prefer CI instead)

# Test Plan

Tested that it correctly errors if package names are passed in.

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [ ] I added a `changelog.md` entry and rebuilt the package sources according to [this short guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)