[CI][expo-updates] Update E2E tests README and comments (#37677)
# Why

Provide instructions for running these tests locally.

# Test Plan

CI should pass. Originally, this PR was removing all the Detox related files and setup code in project.ts. However, doing that seems to break the tests on Android. Further work will be needed to determine which of those files are still needed, and why.