[ios][dev-menu] Fix button press on scroll (#37661)
# Why
Button presses were not cancelled on scroll when using the dev menu.

# How
It was caused by adding a LazyVStack around the scroll view content. Removing it fixes the issue. The view will never be large enough there to require a lazy stack anyway.

# Test Plan
Bare-expo. Button presses are cancelled on scroll