[CI][expo-updates] Remove unused Detox files (#37751)
# Why

After #37707 it should be safe to remove all the unused Detox code and files from updates E2E.

# Test plan

CI should pass.

It turns out we still need the `@config-plugins/detox` plugin to set the Android network security correctly to allow it to reach the test updates server. Since the Detox plugin is no longer maintained, we should probably add our own plugin for customers that want to do this type of testing.