[audio] Fix issue with currentTime when seeking (#37668)
Closes #37653

The time observers are only emitting when the player is playing. This means that if you seek while the player is not playing, the time gets out of sync. To fix it, we should emit a status update when seeking.

Bare-expo. The `currentTime` stays in sync when seeking when the player isn't actively playing.