[video] Export types using the `export type` syntax. (#37396)
# Why

We should export types with `export type` syntax instead of just using
`export`. Additionaly in the video thumbnail export we should be using
`export { default as VideoThumbnails }`.

Doesn't really fix anything for us. There is some project at SWM that
still uses Webpack for web and has issues because of this. We don't
support webpack anymore but we can help them out 😄

# How

Applied the changes described above.

# Test Plan

Tested in BareExpo on web and mobile