[audio] Fix webpack issues (#37421)
# Why

We have marked webpack support as deprecated, one of the projects at SWM
still uses it and for them it raises some errors that we should probably
treat as valid anyways.
- We don't export the types with `export type` annotation
- `ExpoAudio.web.ts` imports from `ExpoAudio`. On webpack this resolves
to `ExpoAudio.web.ts` and fails to import the event names.

# How

Moved the event names into a separate file. Added `type` to the exports

# Test Plan

Tested on web, ran `yarn tsc` for ncl in and `et cp` for the package.
Docs are still looking good as well.

# Checklist