[video][ios] Add AirPlay cast support (#37207)
# Why

Currently the users can cast to AirPlay only using the default controls
enabled with `useNativeControls=true`. We want to make it possible to do
it in custom UIs as well.

# How

Since `AVPlayer` and `AVPlayerViewController` don't provide any APIs for
casting, I've added a new view wrapping `AVRoutePickerView` which
provides functionality to select a device and cast to it.
In case you wonder about it during the review - this view is kind of
magic ✨ , because it selects a relevant player to cast by itself,
without our input - that's why we don't pass a player into it.

Along with the view I've added the `isExternalPlaybackActive` property
and relevant events for monitoring it's state.

# Test Plan

Tested in Bare Expo on iOS 18 device with my macbook as the streaming
target. Also tested on Android to make sure that no errors are thrown.