[video] Add fullscreen orientation and auto-exit functionality (#36910)
# Why

There is currently no way to enter lock the orientation of the video in
fullscreen mode. While I was adding this functionality I realised that
there is no way do add auto-exit functionality from the JS side (android
fullscreen is a separate activity) so I also added a prop that enables
that.

**Note: If you are testing this PR on iOS, use physical device, the iOS
emulator has issues when exiting the fullscreen programatically**

# How

Added `fullscreenOptions` prop that encapsulates all fullscreen
settings. In there we can lock the fullscreen orientation, also users
can enable youtube-like exit of fullscreen when the device is rotated to
a different orientation (for example they were watching a landscape
video and rotated the device to portrait)

# Test Plan

Tested in BareExpo on Android and iOS