Revert "[notifications] remove expo-notifications from auto plugins" (#37886)
This reverts commit e5384b17331531b49e8b811be45104db3e1ddcbd.

this PR is the same as
https://github.com/expo/expo/commit/a5dfa17e53da7802d1bc7d821ffcebdb4659577e
which was picked to sdk-53 branch, but this one is targeting main so
it's included in SDK 54.

We will land the original diff again for SDK 55. Use the time before
that to migrate to `expo-notifications` config plugin if you want to
customize the notification icon or other options.

I'll add a warning for consumers of these options in a follow-up PR, if
agreed.

# Why

This field isn't really documented in our notification guides, some of
the options have no effect, and the `expo-notifications` config plugin
is the go-to way to configure these things. That's why we'll remove it.

But also, some people depend on it so there'll be a deprecation period.

# How

- revert

# Test Plan

green CI

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>