chore(expo-modules-core): restructure jest polyfill setup (#37588)
# Why

- Split out of more complicated PR to add package exports support
https://github.com/expo/expo/pull/37302
- The jest-expo setup script pulls in the "web-only" polyfills for all
platforms to ensure certain globals are setup. This PR renames these
polyfills to not be web-specific since they're really just about
compatibility with the native runtime.
- This PR also refactors the code so it's clearly listing the polyfills
as a dangerous module that shouldn't be imported unless you know what
you're doing.
- Finally, we get rid of the wonky import that uses a platform extension
directly. This will ensure that any changes to jest-expo in the future
won't cause the tests to break in an unexpected way.
- This represents the majority of the changes needed to enable package
exports in expo-modules-core but it does it in a way that hopefully
won't be reverted due to unrelated tests failing.

# Test Plan

- Tests should continue to pass. However, changing jest-expo /
expo-modules-core may lead to unexpected changes across packages that
may not run in CI.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>