[video][android] Fix duration property resetting to 0 on video repeat (#37984)
# Why

Fixes [#34700](https://github.com/expo/expo/issues/34700)

# How
This situation occurs because on repeat the player remains in
`STATE_READY`, and the duration is not refreshed after being set to 0 in
onMediaTransition.
Removed resetting the duration on `MEDIA_ITEM_TRANSITION_REASON_REPEAT`
because the duration does not change on video repeat.

# Test Plan

Tested on bare-expo using the code from the issue.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>