refactor(babel-preset-expo): Re-remove `@babel/core` dependency in `babel-preset-expo` by using host types/template (#38177)
This is based on #38171 which switches as many imports as possible to
`@babel/core`. Furthermore, that meant that the only realised dependency
for Babel packages (apart from sub-plugins/presets/helpers) was
`@babel/core` in `babel-preset-expo`. The PR then added `@babel/core` as
a dependency to `babel-preset-expo`.

This PR refactors the accidentally used imports from `@babel/core`
further to remove them and intead use the host `@babel/core` values
(i.e. what the plugin receives as an argument)
This has slipped through in a few places, and the types were a little
misaligned.