[image][iOS] Fix caching resized images from Photo Library (#38105)
# Why

When loading image from PH with target size like 50x50 it's being cached
as the original image, later, when we want to load the same image with
bigger size, like 1000x1000 its being picked from cache ignoring that
it's bigger size now.

# How

Images from PH are already loaded with `targetSize` so we dont want to
store them as "original" image, instead we store them only as a
processed image with size included.