[font] throw when loading empty font file (#38229)
# Why

This PR adds a check to prevent loading empty font files on Android. This normally wouldn't happen but if it did, Android TypeFace throws `ArrayIndexOutOfBoundsException` which is pretty useless for debugging.

# How

Added a simple check in the `FontLoaderModule.kt` file that verifies if the font file has a non-zero length before attempting to create a Typeface from it. If the file is empty (length is 0), it throws a `CodedException` with a error message that's better than `ArrayIndexOutOfBoundsException`.

# Test Plan

- bare expo

# Checklist

- [x] I added a `changelog.md` entry and rebuilt the package sources according to [this short guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)