[android][updates] Fix procedures across reloads (#38073)
# Why
Fixes an issue where after a reload, you cannot reload a second time. This is because the procedureScope was closed when the updates module is destroyed.

# How
We need the scope to exist for the lifetime of the app so I've removed the call.

# Test Plan
e2e app. Consecutive reloads work as expected.