[updates] Introduce ReloadingView (#38074)
# Why
Closes ENG-15542
Add support for a customizable `ReloadingView` that is displayed after a call to `reloadAsync`

# How
Added new api to the updates module. 
Users can use:
`setReloadScreenOptions` to customize how the screen looks
`show/hide` are useful for debugging how the screen looks. 

We will call `show` during the relaunch procedure and use a similar method that `expo-splash-screen` uses to detect when the app has loaded to hide it again. On Android, we use `ReactMarker` and on iOS we listen for the `RCTContentDidAppearNotification`. 

# Test Plan
Added new screen to ncl and also tested in a e2e project. Working as expected

https://github.com/user-attachments/assets/28cf4f67-0716-4b36-b1f0-fb51a93dcaae


https://github.com/user-attachments/assets/fb409e16-4d07-4965-bd18-7b06723bb870