[dev-menu][android] Add FAB to pull up the dev menu (#38246)
# Why

Adds a Floating Action Button that shows the dev menu. This is mostly
necessary for Oculus Quest support, but also makes development easier on
Android phones.

# How

Added a custom FAB to the view hierarchy. It uses spring-based
animations and snaps to one of the sides of the phone based on movement
speed and current position.

# Test Plan

Tested in BareExpo on Android and Quest



https://github.com/user-attachments/assets/8b91c734-ecb9-49ec-9dff-93216e8917b4