[docs] Clarify "Workflows" under the Hosting section, general capitalization and grammar
Why/How
===
I found "Workflows" in the Hosting section to be ambiguous. This PR changes it to "Deploy with EAS Workflows" and clarifies in the page title that it's about web deployments. For grammatical consistency I changed "Monitoring API routes" to "Monitor API routes" which matches the tense we use in other sections.

Also fixed up various errors throughout the docs.