[docs] Clarify "Workflows" under the Hosting section, note that GitHub integration is optional, general capitalization and grammar (#38316)
Why/How
===
I found "Workflows" in the Hosting section to be ambiguous. This PR
changes it to "Deploy with EAS Workflows" and clarifies in the page
title that it's about web deployments. For grammatical consistency I
changed "Monitoring API routes" to "Monitor API routes" which matches
the tense we use in other sections.

Also, GitHub integration is now optional for Workflows. I changed the
wording of that requirement in automating-eas-cli.mdx.

Also fixed up various errors throughout the docs.