[core][iOS] Fix null string conversion (#38284)
# Why

Converting NSString to JavaScriptValue strips everything after null even
if it's still a valid javascript string.
This issue also exists in react-native:
facebook/react-native/issues/24129

# How

https://github.com/facebook/hermes/blob/8b384559583d98879a7987141d2c430419b8340d/lib/VM/StringPrimitive.cpp#L93

This implementation is essentially the same as what hermes does under
the hood. When creating a `jsi::String` from a utf8 string, hermes
checks if the string `isAllASCII`. If so, it uses `createFromAscii`.
However, null character passes this check, so any content after the null
character is stripped.

To avoid this behavior, we can treat the string as utf16 and create the
`jsi::String` using `createFromUtf16`, which skips the ASCII check and
preserves the entire string, including content after the null character.

* Using `cStringUsingEncoding` should be fine as long as we provide the
`length`.
* `jsi::String::createFromUtf16` was introduced in `react-native@0.79`
but macos is still on `0.78` so it uses a previous implementation until
it's updated.